refactor: remove maa-run and use dlopen to load MaaCore
- remove maa-run and use dlopen to load MaaCore;
- search library and resource in data dir, current dir and relative dir;
