+++ .github/workflows/ci.yml
@@ -52,14 +52,6 @@ jobs:
         run: |
           rustup target add $CARGO_BUILD_TARGET
           rustup component add clippy rustfmt
-      - name: Install MaaCore with maa-updater
-        env:
-          MAA_API_URL: https://github.com/MaaAssistantArknights/MaaRelease/raw/main/MaaAssistantArknights/api/version
-        run: |
-          cargo build --locked --bin maa
-          cargo run --bin maa -- install stable
-          ls -l "$(cargo run --bin maa -- dir library)"
-          ls -l "$(cargo run --bin maa -- dir resource)"
       - name: Build
         run: cargo build --workspace --locked
       - name: Lint (clippy)
@@ -68,12 +60,13 @@ jobs:
         run: cargo fmt --all --check
       - name: Test
         run: cargo test --all-targets --all-features --locked
-      - name: Try to run maa-run
-        run: ./target/$CARGO_BUILD_TARGET/debug/maa-run version
-      - name: Try to run relocated lib with maa
+      - name: Install MaaCore
+        env:
+          MAA_API_URL: https://github.com/MaaAssistantArknights/MaaRelease/raw/main/MaaAssistantArknights/api/version
         run: |
-          maa="./target/$CARGO_BUILD_TARGET/debug/maa"
-          old_data_dir=$($maa dir data)
-          new_data_dir="$PWD/data"
-          mv "$old_data_dir" "$new_data_dir"
-          MAA_DATA_DIR="$new_data_dir" $maa version all
+          cargo run install beta
+          ls -l "$(cargo run --bin maa -- dir library)"
+          ls -l "$(cargo run --bin maa -- dir resource)"
+      - name: Try run
+        run: |
+          cargo run -- version
