+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["maa-cli", "maa-run", "maa-sys"]
+members = ["maa-cli", "maa-sys"]
 
 [workspace.package]
 license = "AGPL-3.0-or-later"
