+++ maa-cli/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "maa-cli"
 authors = ["Loong Wang <wangl-cc@outlook.com>"]
-version = "0.2.3"
+version = "0.3.0"
 edition = "2021"
 license.workspace = true
 
@@ -10,11 +10,15 @@ name = "maa"
 path = "src/main.rs"
 
 [dependencies]
+maa-sys = { path = "../maa-sys", features = ["runtime"] }
 directories = "5"
 anyhow = "1"
 clap = { version = "4.3", features = ["derive"] }
 paste = "1"
 serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+toml = "0.7"
+serde_yaml = "0.9.25"
 reqwest = { version = "0.11", features = ["blocking", "stream", "json"] }
 indicatif = "0.17.6"
 tokio = { version = "1.31", default-features = false, features = ["rt"] }
@@ -25,3 +29,8 @@ zip = { version = "0.6.6", default-features = false, features = ["deflate"] }
 semver = { version = "1.0.18", features = ["serde"] }
 sha2 = "0.10.7"
 digest = "0.10.7"
+
+[dependencies.chrono]
+version = "0.4.26"
+default-features = false
+features = ["std", "clock", "serde"]
