+++ maa-cli/src/log.rs
