--- maa-run/Cargo.toml
@@ -1,25 +1,0 @@
-[package]
-name = "maa-run"
-authors = ["Loong Wang <wangl-cc@outlook.com>"]
-version = "0.2.6"
-edition = "2021"
-license.workspace = true
-
-[dependencies]
-maa-sys = { path = "../maa-sys" }
-clap = { version = "4.3", features = ["derive"] }
-serde = { version = "1", features = ["derive"] }
-directories = "5"
-serde_json = "1"
-anyhow = "1"
-toml = "0.7"
-paste = "1"
-serde_yaml = "0.9.25"
-
-[dependencies.chrono]
-version = "0.4.26"
-default-features = false
-features = ["std", "clock", "serde"]
-
-[build-dependencies]
-directories = "5"
