+++ maa-sys/Cargo.toml
@@ -5,5 +5,11 @@ version = "0.1.1"
 edition = "2021"
 license.workspace = true
 
+[features]
+runtime = ["libloading"]
+
+[dependencies]
+libloading = { version = "0.8.0", optional = true }
+
 [build-dependencies]
 directories = "5"
