+++ maa-sys/build.rs
@@ -1,34 +1,23 @@
-use directories::ProjectDirs;
-use std::env::{
-    consts::{DLL_PREFIX, DLL_SUFFIX},
-    var_os,
-};
-use std::path::PathBuf;
+#[cfg(not(feature = "runtime"))]
+fn static_link() {
+    use std::env::{
+        consts::{DLL_PREFIX, DLL_SUFFIX},
+        var_os,
+    };
+    use std::path::PathBuf;
 
-fn get_data_dir(proj: Option<ProjectDirs>) -> PathBuf {
-    if let Some(maa_dir) = var_os("MAA_DATA_DIR") {
-        PathBuf::from(maa_dir)
-    } else if let Some(xdg_dir) = var_os("XDG_DATA_HOME") {
-        PathBuf::from(xdg_dir).join("maa")
-    } else if let Some(dirs) = proj {
-        dirs.data_dir().to_path_buf()
-    } else {
-        panic!("Failed to get data directory!")
+    let core_dir = var_os("MAA_CORE_DIR")
+        .map(PathBuf::from)
+        .expect("MAA_CORE_DIR not set");
+    let core_name = format!("{}MaaCore{}", DLL_PREFIX, DLL_SUFFIX);
+    if !core_dir.join(core_name).exists() {
+        panic!("cannot find maa core, make sure you have installed maa core at correct path");
     }
+    println!("cargo:rustc-link-search=native={}", core_dir.display());
+    println!("cargo:rustc-link-arg=-Wl,-rpath,{}", core_dir.display());
 }
 
 fn main() {
-    let proj = ProjectDirs::from("com", "loong", "maa");
-    let data_dir = get_data_dir(proj);
-    let lib_dir = data_dir.join("lib");
-    let core_name = format!("{}MaaCore{}", DLL_PREFIX, DLL_SUFFIX);
-    if !lib_dir.join(core_name).exists() {
-        panic!("cannot find maa core, make sure you have installed maa core at correct path");
-    }
-    println!("cargo:rustc-link-search=native={}", lib_dir.display());
-    if cfg!(target_os = "windows") {
-        println!("cargo:rustc-link-arg=/LIBPATH:{}", lib_dir.display());
-    } else {
-        println!("cargo:rustc-link-arg=-Wl,-rpath,{}", lib_dir.display());
-    }
+    #[cfg(not(feature = "runtime"))]
+    static_link();
 }
